+++ driver/src/test/functional/com/mongodb/reactivestreams/client/SmokeTestSpecification.groovy
@@ -17,6 +17,7 @@
 package com.mongodb.reactivestreams.client
 
 import com.mongodb.MongoNamespace
+import com.mongodb.client.MongoDriverInformation
 import com.mongodb.client.model.IndexModel
 import com.mongodb.diagnostics.logging.Loggers
 import org.bson.Document
@@ -152,6 +153,17 @@ class SmokeTestSpecification extends FunctionalSpecification {
 
         then:
         thrown(IllegalStateException)
+    }
+
+    def 'should accept custom MongoDriverInformation'() {
+        when:
+        def driverInformation = MongoDriverInformation.builder().driverName('test').driverVersion('1.2.0').build()
+
+        then:
+        def client = MongoClients.create(getConnectionString(), driverInformation)
+
+        cleanup:
+        client?.close()
     }
 
     @SuppressWarnings('BusyWait')