+++ docs/reference/content/changelog.md
@@ -15,6 +15,7 @@ Changes between released versions
 [Full change list](https://jira.mongodb.org/issues/?jql=project%20%3D%20JAVARS%20AND%20fixVersion%20%3D%201.3)
 
   * Updated MongoDB Driver Async to 3.4.0-beta1
+  * Added support for views [JAVARS-22](https://jira.mongodb.org/browse/JAVARS-22)
 
 ### 1.2