+++ docs/reference/content/changelog.md
@@ -10,6 +10,12 @@ title = "Changelog"
 
 Changes between released versions
 
+### 1.4
+[Full change list](https://jira.mongodb.org/issues/?jql=project%20%3D%20JAVARS%20AND%20fixVersion%20%3D%201.4)
+
+  * Updated MongoDB Driver Async to 3.4.2
+  * Added GridFS support [JAVARS-21](https://jira.mongodb.org/browse/JAVARS-21)
+
 ### 1.3
 
 [Full change list](https://jira.mongodb.org/issues/?jql=project%20%3D%20JAVARS%20AND%20fixVersion%20%3D%201.3)